Move controller theming code to concern
